+++ C/include/heap_sort.h
@@ -1,5 +1,7 @@
 #pragma once
 
+#include <stdlib.h>
+
 typedef int (*HSCompareFun)(const void*, const void*);
 
-void heap_sort(void** array, unsigned int count, HSCompareFun compare);
+void heap_sort(void** array, size_t count, HSCompareFun compare);
