+++ C/src/merge_sort.c
@@ -2,11 +2,11 @@
 #include <stdlib.h>
 #include "array_g.h"
 
-static void merge(void** array, unsigned int start, unsigned int mid, unsigned int end, MSCompareFun compare) {
+static void merge(void** array, size_t start, size_t mid, size_t end, MSCompareFun compare) {
   void** buf = (void**) malloc(sizeof(const void*)*(end-start+1));
-  unsigned int i = start;
-  unsigned int j = mid+1;
-  unsigned int k = 0;
+  size_t i = start;
+  size_t j = mid+1;
+  size_t k = 0;
 
   while(i<=mid && j<=end) {
     if(compare(array[i], array[j]) <= 0) {
@@ -32,11 +32,11 @@ static void merge(void** array, unsigned int start, unsigned int mid, unsigned i
   free(buf);
 }
 
-static void merge_g(void* array, unsigned int start, unsigned int mid, unsigned int end, unsigned int size, MSCompareFun compare) {
+static void merge_g(void* array, size_t start, size_t mid, size_t end, size_t size, MSCompareFun compare) {
   void* buf = (void*) malloc(size*(end-start+1));
-  unsigned int i = start;
-  unsigned int j = mid+1;
-  unsigned int k = 0;
+  size_t i = start;
+  size_t j = mid+1;
+  size_t k = 0;
 
   while(i<=mid && j<=end) {
     if(compare(at_g(array, i, size), at_g(array,j, size)) <= 0) {
@@ -62,22 +62,22 @@ static void merge_g(void* array, unsigned int start, unsigned int mid, unsigned 
   free(buf);
 }
 
-static void merge_sort_(void** array, unsigned int start, unsigned int end, MSCompareFun compare) {
+static void merge_sort_(void** array, size_t start, size_t end, MSCompareFun compare) {
   if(start >= end)
     return;
 
-  unsigned int mid = (start + end)/2;
+  size_t mid = (start + end)/2;
 
   merge_sort_(array, start, mid, compare);
   merge_sort_(array, mid+1, end, compare);
   merge(array, start, mid, end, compare);
 }
 
-static void merge_sort_g_(void* array, unsigned int start, unsigned int end, unsigned int size, MSCompareFun compare) {
+static void merge_sort_g_(void* array, size_t start, size_t end, size_t size, MSCompareFun compare) {
   if(start >= end)
     return;
 
-  unsigned int mid = (start + end)/2;
+  size_t mid = (start + end)/2;
 
   merge_sort_g_(array, start, mid, size, compare);
   merge_sort_g_(array, mid+1, end, size, compare);
@@ -85,14 +85,14 @@ static void merge_sort_g_(void* array, unsigned int start, unsigned int end, uns
 }
 
 
-void merge_sort_g(void* array, unsigned int count, unsigned int size, MSCompareFun compare) {
+void merge_sort_g(void* array, size_t count, size_t size, MSCompareFun compare) {
   if(count==0) {
     return;
   }
   merge_sort_g_(array, 0, count-1, size, compare);
 }
 
-void merge_sort(void** array, unsigned int count, MSCompareFun compare) {
+void merge_sort(void** array, size_t count, MSCompareFun compare) {
   if(count == 0) {
     return;
   }
